+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.2] - 2026-01-11
+
+### Added
+- `time_to_datetime()` utility function to convert Skyfield Time to Python datetime
+- `is_planet_in_retrograde()` function to get the tuple `(bool, start_date, end_date)` for planet retrograde.
+
 ## [0.4.1] - 2026-01-09
 
 ### Changed